Changed Blind Xfer to utilize existing logic for checking where to dial & if you can dial.
NOTE: Blind Xfer checks if the Caller can dial if the Destination connection wants to blind xfer the Caller connection. If the Caller wants to blind xfer the Destination connection, we set the ani to the organization number (hardcoded into the PBX app) so it can bypass the IP check.

git-svn-id: svn+ssh://cvs1/svn/listen/trunk@3242 4692729f-32a0-46b7-9832-1bf53c4881b5